+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `prefix` parameter to the unimock macro, in case the crate is re-exported through another crate.
 
 ## [0.3.6] - 2022-07-11
 ### Changed